+++ nirc_ehr/resources/views/begin.html
@@ -154,6 +154,7 @@
                         },
                         items: [
                             {name: 'Active Clinical Cases', url: '<%=contextPath%>' + ctx['EHRStudyContainer'] + '/ehr-animalHistory.view#inputType:none&showReport:1&activeReport:cases'},
+                            {name: 'Active Behavior Cases', url: '<%=contextPath%>' + ctx['EHRStudyContainer'] + '/ehr-animalHistory.view#inputType:none&showReport:1&activeReport:behaviorCases'},
                           ]
                     },
                     {
