+++ nirc_ehr/resources/web/nirc_ehr/model/sources/BehavioralCase.js
@@ -107,54 +107,6 @@ EHR.model.DataModelManager.registerMetadata('BehavioralCase', {
             formCasesLink: {
                 xtype: 'nirc_ehr-editCases'
             }
-        },
-        'study.clinremarks': {
-            hx: {
-                formEditorConfig: {
-                    xtype: 'ehr-hxtextarea'
-                },
-            },
-            p: {
-                height: 120
-            },
-            remark: {
-                label: 'Remark',
-                height: 120
-            },
-            vetreview: {
-                height: 120
-            },
-            category: {
-                getInitialValue: function (v, rec) {
-                    return 'Behavior'
-                },
-                editable: false,
-                hidden: true,
-                columnConfig: {
-                    editable: false
-                }
-            },
-            performedby: {
-                hidden: false,
-                defaultValue: LABKEY.Security.currentUser.id.toString(),
-                editorConfig: {
-                    store: {
-                        type: 'labkey-store',
-                        schemaName: 'core',
-                        sql: 'SELECT CAST (UserId AS VARCHAR) AS UserId,DisplayName,FirstName,LastName FROM core.users',
-                        autoLoad: true
-                    }
-                }
-            },
-            enddate: {
-                hidden: true
-            },
-            dateFinalized: {
-                hidden: true
-            },
-            qcstate: {
-                hidden: true
-            },
-        },
+        }
     }
 });
